backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy[asyncio]==2.0.27
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
google-generativeai==0.3.2
pytest==8.0.0
httpx==0.26.0
python-dotenv==1.0.1
email-validator==2.1.0.post1
bcrypt==3.2.0
